Change MONGODB_URI environment variable to DB_URI [#175291468]
As part of the mLabs to Atlas migration the Mongo DB config var needs to be renamed from MONGODB_URI as it will be auto removed when the mLabs add-on is deleted.  After this removal the app will be restarted with the DB_URI set in the Heroku settings before the restart.